Add a model mapper to map model to provider strings for LiteLLM
