+++ libs/superagent/app/completion/types.py
@@ -1,0 +1,5 @@
+MODEL_MAPPER = {
+    "HUGGINGFACE": {
+        "MISTRAL_7B_INSTRUCT_V01": "huggingface/mistralai/Mistral-7B-Instruct-v0.1"
+    }
+}
